separate Queue.h from Queue.cc
